refactor views and clean up model
