+++ Apple-Frameworks/FrameworkDetailView.swift
@@ -17,17 +17,7 @@ struct FrameworkDetailView: View {
         
         VStack {
             
-            HStack {
-                Spacer()
-                Button{
-                    isShowingDetailView = false
-                } label: {
-                    Image(systemName: "xmark")
-                        .foregroundColor(Color(.label))
-                        .imageScale(.large)
-                        .frame(width: 44, height: 44, alignment: /*@START_MENU_TOKEN@*/.center/*@END_MENU_TOKEN@*/)
-            }
-            }.padding()
+            xDismissButton(isShowingDetailView: $isShowingDetailView)
             
             Spacer()
             FrameworkTitleView(framework: framework)
