+++ Apple-Frameworks/FrameworkGridViewModel.swift
@@ -8,10 +8,14 @@
 import SwiftUI
 
 final class FrameworkGridViewModel: ObservableObject {
+    
+    let columns: [GridItem] = [GridItem(.flexible()),
+                               GridItem(.flexible()),
+                               GridItem(.flexible())
+    ]
+    
     var selectedFramework: Framework? {
-        didSet {
-            isShowingDetailView = true
-        }
+        didSet { isShowingDetailView = true }
     }
