+++ Apple-Frameworks/FrameworkTitleView.swift
@@ -1,0 +1,35 @@
+//
+//  FrameworkTitleView.swift
+//  Apple-Frameworks
+//
+//  Created by Mark Martin on 7/7/21.
+//
+
+import SwiftUI
+
+struct FrameworkTitleView : View {
+    
+    let framework : Framework
+    
+    var body: some View {
+        VStack {
+            Image(framework.imageName)
+                .resizable()
+                .frame(width: 90, height: 90, alignment: /*@START_MENU_TOKEN@*/.center/*@END_MENU_TOKEN@*/)
+            Text(framework.name)
+                .font(.title2)
+                .fontWeight(.semibold)
+                .scaledToFit()
+                .minimumScaleFactor(0.6)
+        }
+        .padding()
+    }
+    
+    
+}
+
+struct FrameworkTitleView_Previews: PreviewProvider {
+    static var previews: some View {
+        FrameworkTitleView(framework: MockData.sampleFramework)
+    }
+}
